+++ app/static/lib/attribute-classes.js
@@ -2,11 +2,17 @@
 
 export const attCurvature = ['bend', 'curve', 'lv', 'phrase', 'slur', 'tie'];
 
+// see att.placement in MEI 4.0.1 definition: 'measure' taken out
 export const attPlacement = [
   'accid', 'artic', 'attacca', 'breath', 'caesura', 'cpMark', 'dir', 'dynam',
   'dynam', 'f', 'fermata', 'fing', 'fingGrp', 'hairpin', 'harm', 'harpPedal',
-  'lg', 'line', 'measure', 'metaMark', 'mNum', 'mordent', 'ornam', 'pedal',
+  'lg', 'line', 'metaMark', 'mNum', 'mordent', 'ornam', 'pedal',
   'refrain', 'reh', 'sp', 'stageDir', 'syl', 'tempo', 'trill', 'turn', 'verse'
+];
+
+// pedal not yet in vertical group, but supported by Verovio; attacca, tempo not supported
+export const attVerticalGroup = [
+  'attacca', 'dir', 'dynam', 'hairpin', 'tempo', 'pedal'
 ];
 
 export const dataPlacement = ['above', 'within', 'between', 'below'];