+++ app/static/lib/editor.js
@@ -19,7 +19,7 @@ export function delEl(v, cm) {
   let selectedElements = [];
   v.updateNotation = false;
   // let checkPoint = buffer.createCheckpoint(); TODO
-  if (att.modelControlEvents.concat(['accid', 'artic', 'clef', 'octave'])
+  if (att.modelControlEvents.concat(['accid', 'artic', 'clef', 'octave', 'beamSpan'])
     .includes(element.nodeName)) {
     if (element.nodeName == 'octave') { // reset notes inside octave range
       let disPlace = element.getAttribute('dis.place');
@@ -222,7 +222,7 @@ export function invertPlacement(v, cm, modifier = false) {
   for (let id of ids) {
     var el = v.xmlDoc.querySelector("[*|id='" + id + "']");
     let chordId = utils.insideParent(id);
-    if (el && el.nodeName == 'note') {
+    if (el && el.nodeName === 'note') {
       if (chordId) id = chordId;
       el = v.xmlDoc.querySelector("[*|id='" + id + "']");
     }
@@ -233,13 +233,13 @@ export function invertPlacement(v, cm, modifier = false) {
     let attr = '';
     let val = 'above';
     // placement above/below as in dir, dynam...
-    if (att.attPlacement.includes(el.nodeName)) {
+    if (att.attPlacement.includes(el.nodeName)) { //|| el.nodeName === 'beamSpan') {
       attr = 'place';
       if (el.hasAttribute(attr) &&
         (att.dataPlacement.includes(val) && el.getAttribute(attr) != 'below')) {
         val = 'below';
       }
-      if (el.nodeName == 'fermata')
+      if (el.nodeName === 'fermata')
         val == 'below' ?
         el.setAttribute('form', 'inv') : el.removeAttribute('form');
       el.setAttribute(attr, val);
@@ -262,7 +262,7 @@ export function invertPlacement(v, cm, modifier = false) {
       range = replaceInTextEditor(cm, el, true);
       // txtEdr.autoIndentSelectedRows();
       // invert @num.place within tuplet
-    } else if (el.nodeName == 'tuplet') {
+    } else if (el.nodeName === 'tuplet') {
       attr = 'num.place';
       val = 'above';
       if (el.hasAttribute(attr) && el.getAttribute(attr) == val) {
@@ -271,13 +271,30 @@ export function invertPlacement(v, cm, modifier = false) {
       el.setAttribute(attr, val);
       range = replaceInTextEditor(cm, el, true);
       // txtEdr.autoIndentSelectedRows();
+    } else if (el.nodeName === 'beamSpan') { // replace individual notes in beamSpan
+      attr = 'stem.dir', val = 'up';
+      let plist = el.getAttribute('plist');
+      if (plist) {
+        plist.split(' ').forEach(p => {
+          let note = v.xmlDoc.querySelector("[*|id='" + speed.rmHash(p) + "']");
+          if (note) {
+            if (note.parentNode.nodeName === 'chord') note = note.parentNode;
+            if (note.hasAttribute(attr) && note.getAttribute(attr) == val) {
+              val = 'down';
+            }
+            note.setAttribute(attr, val);
+          }
+          v.updateNotation = false; // no need to redraw notation
+          range = replaceInTextEditor(cm, note, true);
+        });
+      }
       // find all note/chord elements children and execute InvertingAction
     } else if (noteList = el.querySelectorAll("note, chord")) {
       // console.info('noteList: ', noteList);
       attr = 'stem.dir', val = 'up';
       for (let note of noteList) {
         // skip notes within chords
-        if (note.parentNode.nodeName == 'chord') continue;
+        if (note.parentNode.nodeName === 'chord') continue;
         if (note.hasAttribute(attr) && note.getAttribute(attr) == val) {
           val = 'down';
         }
@@ -438,6 +455,38 @@ export function addBeamElement(v, cm, elementName = 'beam') {
   v.updateNotation = true; // update notation again
 } // addBeamElement()
 
+// add beamSpan element
+export function addBeamSpan(v, cm) {
+  v.loadXml(cm.getValue());
+  if (v.selectedElements.length < 1) return;
+  v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
+  let id1 = v.selectedElements[0]; // xml:id string
+  let id2 = v.selectedElements[v.selectedElements.length - 1];
+  // add control like element <octave @startid @endid @dis @dis.place>
+  let beamSpan = v.xmlDoc.createElementNS(speed.meiNameSpace, 'beamSpan');
+  let uuid = 'beamSpan-' + utils.generateUUID();
+  beamSpan.setAttributeNS(speed.xmlNameSpace, 'xml:id', uuid);
+  beamSpan.setAttribute('startid', '#' + id1);
+  beamSpan.setAttribute('endid', '#' + id2);
+  beamSpan.setAttribute('plist', v.selectedElements.map(e => '#' + e).join(' '));
+  let n1 = v.xmlDoc.querySelector("[*|id='" + id1 + "']");
+  n1.closest('measure').appendChild(beamSpan);
+  v.updateNotation = false;
+  let sc = cm.getSearchCursor('xml:id="' + id1 + '"');
+  if (sc.findNext()) {
+    let p1 = utils.moveCursorToEndOfMeasure(cm, sc.from());
+    cm.replaceRange(speed.xmlToString(beamSpan) + '\n', cm.getCursor());
+    cm.indentLine(p1.line, 'smart'); // TODO
+    cm.indentLine(p1.line + 1, 'smart');
+    cm.setSelection(p1);
+  }
+  v.selectedElements = [];
+  v.selectedElements.push(uuid);
+  v.lastNoteId = id2;
+  v.updateData(cm, false, true);
+  v.updateNotation = true; // update notation again
+} // addBeamSpan()
+
 // add octave element and modify notes inside selected elements
 export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
   v.loadXml(cm.getValue());
@@ -480,6 +529,70 @@ export function addOctaveElement(v, cm, disPlace = 'above', dis = '8') {
   v.updateData(cm, false, true);
   v.updateNotation = true; // update notation again
 } // addOctaveElement()
+
+// surround selected elements with a supplied element (and a responsibility
+// statement from v.respId
+export function addSuppliedElement(v, cm) {
+  v.loadXml(cm.getValue());
+  v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
+  v.selectedElements = utils.sortElementsByScorePosition(v.selectedElements);
+  if (v.selectedElements.length < 1) return;
+  v.updateNotation = false;
+  let uuids = [];
+  v.selectedElements.forEach(id => {
+    let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
+    if (!el) {
+      console.warn('No such element in xml document: ' + id);
+    } else {
+      let parent = el.parentNode;
+      let sup = document.createElementNS(speed.meiNameSpace, 'supplied');
+      let uuid = 'supplied-' + utils.generateUUID();
+      sup.setAttributeNS(speed.xmlNameSpace, 'xml:id', uuid);
+      if (v.respId) sup.setAttributeNS(speed.xmlNameSpace, 'resp', '#' + v.respId);
+      parent.replaceChild(sup, el);
+      sup.appendChild(el);
+      replaceInTextEditor(cm, el, true, sup);
+      cm.execCommand('indentAuto');
+      uuids.push(uuid);
+    }
+  });
+  // buffer.groupChangesSinceCheckpoint(checkPoint); // TODO
+  v.selectedElements = [];
+  uuids.forEach(u => v.selectedElements.push(u));
+  v.updateData(cm, false, true);
+  v.updateNotation = true; // update notation again
+} // addSuppliedElement()
+
+export function addVerticalGroup(v, cm) {
+  v.loadXml(cm.getValue());
+  v.selectedElements = speed.filterElements(v.selectedElements, v.xmlDoc);
+  if (v.selectedElements.length < 1) return;
+  v.updateNotation = false;
+  let value = 1;
+  let existingValues = []; // search for existing vgrp values on SVG page
+  // look to current page SVG dynam@vgrp, dir@vgrp, hairpin@vgrp, pedal@vgrp
+  // and increment value if already taken
+  document.querySelectorAll('g[data-vgrp]').forEach(e => {
+    let value = parseInt(e.getAttribute('data-vgrp'));
+    if (existingValues.indexOf(value) < 0) existingValues.push(value);
+  });
+  while (existingValues.indexOf(value) >= 0) value++; // increment until unique
+  v.selectedElements.forEach(id => {
+    let el = v.xmlDoc.querySelector("[*|id='" + id + "']");
+    if (!el) {
+      console.warn('No such element in xml document: ' + id);
+    } else if (att.attVerticalGroup.includes(el.nodeName)) {
+      let oldEl = el.cloneNode(true);
+      el.setAttribute('vgrp', value);
+      replaceInTextEditor(cm, oldEl, true, el);
+      cm.execCommand('indentAuto');
+    } else {
+      console.warn('Vertical group not supported for ', el);
+    }
+  });
+  v.updateData(cm, false, true);
+  v.updateNotation = true; // update notation again
+} // addVerticalGroup()
 
 // wrapper for cleaning superfluous @accid.ges attributes
 export function cleanAccid(v, cm) {