+++ app/static/lib/main.js
@@ -94,6 +94,7 @@ export function loadDataInEditor(mei, setFreshlyLoaded = true) {
   v.loadXml(mei);
   let bs = document.getElementById('breaks-select');
   if (bs) bs.value = v.containsBreaks() ? 'line' : 'auto';
+  v.setRespSelectOptions();
 }
 
 export function updateLocalStorage(meiXml) {
@@ -165,6 +166,9 @@ import {
   navElsSelector,
   getElementAtCursor
 } from './dom-utils.js';
+import {
+  addDragSelector
+} from './drag-selector.js';
 import * as e from './editor.js'
 import Viewer from './viewer.js';
 import Storage from './storage.js';
@@ -181,8 +185,8 @@ import {
 import default_schema from '../schemaInfo/mei-CMN-4.0.1.schemaInfo.js';
 
 // mei-friend version and date
-const version = 'develop-0.3.6';
-const versionDate = '4 March 2022';
+const version = 'develop-0.3.8';
+const versionDate = '15 March 2022';
 // const defaultMeiFileName = `${root}Beethoven_WoOAnh5_Nr1_1-Breitkopf.mei`;
 const defaultMeiFileName = `${root}Beethoven_WoO70-Breitkopf.mei`;
 const defaultVerovioOptions = {
@@ -191,7 +195,8 @@ const defaultVerovioOptions = {
   header: "encoded",
   footer: "encoded",
   inputFrom: "mei",
-  adjustPageHeight: "true",
+  adjustPageHeight: true,
+  mdivAll: true,
   outputIndent: 3,
   pageMarginLeft: 50,
   pageMarginRight: 25,
@@ -201,7 +206,9 @@ const defaultVerovioOptions = {
   spacingNonLinear: .5,
   minLastJustification: 0,
   clefChangeFactor: .83,
-  svgAdditionalAttribute: ["layer@n", "staff@n"],
+  svgAdditionalAttribute: ["layer@n", "staff@n",
+    "dir@vgrp", "dynam@vgrp", "hairpin@vgrp", "pedal@vgrp"
+  ],
   bottomMarginArtic: 1.2,
   topMarginArtic: 1.2
 };
@@ -222,7 +229,6 @@ const defaultCodeMirrorOptions = {
   foldGutter: true,
   gutters: ["CodeMirror-linenumbers", "CodeMirror-foldgutter"],
   extraKeys: {
-    "Alt-F": "findPersistent",
     "'<'": completeAfter,
     "'/'": completeIfAfterLt,
     "' '": completeIfInTag,
@@ -230,11 +236,7 @@ const defaultCodeMirrorOptions = {
     "Ctrl-Space": "autocomplete",
     "Alt-.": consultGuidelines
   },
-  hintOptions: {
-    schemaInfo: {
-      ...default_schema
-    }
-  },
+  hintOptions: 'schema_meiCMN_401', // not cm conform: just provide schema name
   theme: 'default',
   zoomFont: 100, // my own option
   matchTheme: false, // notation matches editor theme (my option)
@@ -482,6 +484,7 @@ function vrvWorkerEventsHandler(ev) {
       tkVersion = ev.data.version;
       tkAvailableOptions = ev.data.availableOptions;
       v.addVrvOptionsToSettingsPanel(tkAvailableOptions, defaultVerovioOptions);
+      v.addMeiFriendOptionsToSettingsPanel();
       document.querySelector(".rightfoot").innerHTML +=
         `&nbsp;<a href="https://www.verovio.org/">Verovio ${tkVersion}</a>.`;
       document.querySelector(".statusbar").innerHTML =
@@ -496,7 +499,7 @@ function vrvWorkerEventsHandler(ev) {
         v.updateNotation = false;
         loadDataInEditor(storage.content);
         v.updateNotation = true;
-        v.updateAll(cm, defaultVerovioOptions);
+        v.updateAll(cm);
       }
       v.busy(false);
       break;
@@ -679,7 +682,8 @@ export function openFile(file = defaultMeiFileName, setFreshlyLoaded = true) {
 }
 
 // checks format of encoding string and imports or loads data/notation
-function handleEncoding(mei, setFreshlyLoaded = true) {
+// mei argument may be MEI or any other supported format (text/binary)
+export function handleEncoding(mei, setFreshlyLoaded = true) {
   let found = false;
   v.clear();
   v.busy();
@@ -830,7 +834,6 @@ let cmd = {
   'previousPage': () => v.updatePage(cm, 'backwards'),
   'nextPage': () => v.updatePage(cm, 'forwards'),
   'lastPage': () => v.updatePage(cm, 'last'),
-  // 'nightMode': () => v.swapNotationColors(),
   'nextNote': () => v.navigate(cm, 'note', 'forwards'),
   'previousNote': () => v.navigate(cm, 'note', 'backwards'),
   'nextMeasure': () => v.navigate(cm, 'measure', 'forwards'),
@@ -891,6 +894,7 @@ let cmd = {
   //
   'delete': () => e.delEl(v, cm),
   'invertPlacement': () => e.invertPlacement(v, cm),
+  'addVerticalGroup': () => e.addVerticalGroup(v, cm),
   'toggleStacc': () => e.toggleArtic(v, cm, 'stacc'),
   'toggleAccent': () => e.toggleArtic(v, cm, 'acc'),
   'toggleTenuto': () => e.toggleArtic(v, cm, 'ten'),
@@ -914,6 +918,8 @@ let cmd = {
   'addCClefChangeAfter': () => e.addClefChange(v, cm, 'C', '3', false),
   'addFClefChangeAfter': () => e.addClefChange(v, cm, 'F', '4', false),
   'addBeam': () => e.addBeamElement(v, cm),
+  'addBeamSpan': () => e.addBeamSpan(v, cm),
+  'addSupplied': () => e.addSuppliedElement(v, cm),
   'cleanAccid': () => e.cleanAccid(v, cm),
   'renumberMeasuresTest': () => e.renumberMeasures(v, cm, false),
   'renumberMeasures': () => e.renumberMeasures(v, cm, true),
@@ -933,11 +939,14 @@ let cmd = {
 
 // add event listeners when controls menu has been instantiated
 function addEventListeners(v, cm) {
+  let vp = document.querySelector('.verovio-panel');
+
   // layout notation position
   document.getElementById('top').addEventListener('click', cmd.notationTop);
   document.getElementById('bottom').addEventListener('click', cmd.notationBottom);
   document.getElementById('left').addEventListener('click', cmd.notationLeft);
   document.getElementById('right').addEventListener('click', cmd.notationRight);
+
   // show settings panel
   document.getElementById('showSettingsMenu').addEventListener('click', cmd.showSettingsPanel);
   document.getElementById('showSettingsButton').addEventListener('click', cmd.showSettingsPanel);
@@ -953,6 +962,15 @@ function addEventListeners(v, cm) {
   document.getElementById('SaveMei').addEventListener('click', downloadMei);
   document.getElementById('SaveSvg').addEventListener('click', downloadSvg);
   document.getElementById('SaveMidi').addEventListener('click', downloadMidi);
+
+  // edit dialogs
+  document.getElementById('startSearch').addEventListener('click', () => CodeMirror.commands.find(cm));
+  document.getElementById('findNext').addEventListener('click', () => CodeMirror.commands.findNext(cm));
+  document.getElementById('findPrevious').addEventListener('click', () => CodeMirror.commands.findPrev(cm));
+  document.getElementById('replace').addEventListener('click', () => CodeMirror.commands.replace(cm));
+  document.getElementById('replaceAll').addEventListener('click', () => CodeMirror.commands.replaceAll(cm));
+  document.getElementById('jumpToLine').addEventListener('click', () => CodeMirror.commands.jumpToLine(cm));
+  document.querySelectorAll('.keyShortCut').forEach(e => e.classList.add(navigator.platform.startsWith('Mac') ? 'platform-mac' : 'platform-nonmac'));
 
   // open URL interface
   document.getElementById('openUrlButton').addEventListener('click', cmd.openUrlFetch);
@@ -971,13 +989,13 @@ function addEventListeners(v, cm) {
   fc.addEventListener("dragstart", (ev) => console.log('Drag Start', ev));
   fc.addEventListener("dragend", (ev) => console.log('Drag End', ev));
 
-  // document.getElementById('notation-night-mode-btn').addEventListener('click', cmd.nightMode);
   // Zooming with buttons
   document.getElementById('decrease-scale-btn').addEventListener('click', cmd.zoomOut);
   document.getElementById('increase-scale-btn').addEventListener('click', cmd.zoomIn);
   document.getElementById('verovio-zoom').addEventListener('click', cmd.zoomSlider);
+
   // Zooming notation with mouse wheel
-  document.querySelector('.verovio-panel').addEventListener('wheel', ev => {
+  vp.addEventListener('wheel', ev => {
     if ((navigator.platform.toLowerCase().startsWith('mac') && ev.metaKey) ||
       !navigator.platform.toLowerCase().startsWith('mac') && ev.ctrlKey) {
       ev.preventDefault();
@@ -1015,6 +1033,7 @@ function addEventListeners(v, cm) {
   document.getElementById('downwards-btn').addEventListener('click', cmd.layerDown);
   // manipulation
   document.getElementById('invertPlacement').addEventListener('click', cmd.invertPlacement);
+  document.getElementById('addVerticalGroup').addEventListener('click', cmd.addVerticalGroup);
   document.getElementById('delete').addEventListener('click', cmd.delete);
   document.getElementById('pitchUp').addEventListener('click', cmd.shiftPitchNameUp);
   document.getElementById('pitchDown').addEventListener('click', cmd.shiftPitchNameDown);
@@ -1038,6 +1057,8 @@ function addEventListeners(v, cm) {
   document.getElementById('addCresHairpin').addEventListener('click', cmd.addCresHairpin);
   document.getElementById('addDimHairpin').addEventListener('click', cmd.addDimHairpin);
   document.getElementById('addBeam').addEventListener('click', cmd.addBeam);
+  document.getElementById('addBeamSpan').addEventListener('click', cmd.addBeamSpan);
+  document.getElementById('addSupplied').addEventListener('click', cmd.addSupplied);
   document.getElementById('addArpeggio').addEventListener('click', cmd.addArpeggio);
   // more control elements
   document.getElementById('addFermata').addEventListener('click', cmd.addFermata);
@@ -1171,6 +1192,8 @@ function addEventListeners(v, cm) {
     //   v.pageBreaks = {};
     v.updateAll(cm, {}, v.selectedElements[0]);
   });
+
+  addDragSelector(v, vp);
 } // addEventListeners()