+++ app/static/lib/utils.js
@@ -52,8 +52,8 @@ export function getElementAttributeAbove(cm, row, elementName = 'staff',
   let line;
   while (line = cm.getLine(--row)) {
     if (line.includes('<' + elementName)) {
-      // col = line.indexOf()
-      return [line.match(searchString)[1], row];
+      let match = line.match(searchString);
+      return [(match && match.length > 0) ? match[1] : 1, row];
     }
   }
   return [null, null];
@@ -66,7 +66,8 @@ export function getElementAttributeBelow(cm, row, elementName = 'staff',
   let line;
   while (line = cm.getLine(++row)) {
     if (line.includes('<' + elementName)) {
-      return [line.match(searchString)[1], row];
+      let match = line.match(searchString);
+      return [(match && match.length > 0) ? match[1] : 1, row];
     }
   }
   return [null, null];
@@ -474,4 +475,31 @@ export function attrAsElements(xmlNote) {
 
 export function getOS() {
   return navigator.userAgentData.platform;
+}
+
+// accepts color as string: "rgb(100,12,255)" and hex string "#ffee10" or
+export function brighter(rgbString, deltaPercent) {
+  let rgb = [];
+  if (rgbString.startsWith('#')) {
+    rgb = hexToRgb(rgbString);
+  } else {
+    rgb = rgbString.slice(4, -1).split(',');
+  }
+  rgb = rgb.map(i => Math.max(0, Math.min(parseInt(i) + deltaPercent, 255)));
+  return 'rgb(' + rgb.join(', ') + ')';
+}
+
+function hexToRgb(hex) {
+  return hex.replace(/^#?([a-f\d])([a-f\d])([a-f\d])$/i,
+      (m, r, g, b) => '#' + r + r + g + g + b + b)
+    .substring(1).match(/.{2}/g)
+    .map(x => parseInt(x, 16))
+}
+
+export function complementary(rgbString) {
+  let rgb = [];
+  rgbString.slice(4, -1).split(',').forEach(i => {
+    rgb.push(255 - i);
+  });
+  return 'rgb(' + rgb.join(', ') + ')';
 }