+++ CHANGELOG.md
@@ -118,3 +118,15 @@
 * mei-friend appearance matches the selected theme
 * Option to enable notation to match editor theme colors
 * Theme changes automatically upon OS appearance change (e.g., from bright to dark) when in default theme
+# 0.3.6 patch: mei-friend settings and new functionality
+* Support for surrounding selected elements with a supplied element
+* Settings to control supplied element display, color, and responsibility statement
+* Support to insert vertical group attributes to selected elements (automatically finding unused vgrp value on page)
+* Support for inserting, inverting and deleting beamSpan elements
+# 0.3.7 patch
+* Address issue with ALT-F interfering with EMACS keymap
+* Support for drag-selecting different element categories with mouse
+* Added edit dropdown menu for search/replace/jump-to-line functionality
+# 0.3.8 patch
+* Refactor writing to github, fixing write-to-subdirectory in the process
+